Fix dependency-graph link in README.
Chose to link to the raw svg because it includes the renaming update and
I could not find a link to the raw png.

I tried to regenerate the depgraph but coqdep didn't like the -dumpgraph
flag, and it does not have -v nor --version, nor is it listed in my opam
switch. Not sure what to do about it.